+++ Services/Rds/V3/Model/ListDbAgentJobHistorysResponse.cs
@@ -19,8 +19,8 @@ namespace HuaweiCloud.SDK.Rds.V3.Model
         /// <summary>
         /// 执行历史列表。
         /// </summary>
-        [JsonProperty("historys", NullValueHandling = NullValueHandling.Ignore)]
-        public List<ListDbAgentJobHistorysResult> Historys { get; set; }
+        [JsonProperty("histories", NullValueHandling = NullValueHandling.Ignore)]
+        public List<ListDbAgentJobHistorysResult> Histories { get; set; }
 
         /// <summary>
         /// 执行历史总数。
@@ -37,7 +37,7 @@ namespace HuaweiCloud.SDK.Rds.V3.Model
         {
             var sb = new StringBuilder();
             sb.Append("class ListDbAgentJobHistorysResponse {\n");
-            sb.Append("  historys: ").Append(Historys).Append("\n");
+            sb.Append("  histories: ").Append(Histories).Append("\n");
             sb.Append("  totalCount: ").Append(TotalCount).Append("\n");
             sb.Append("}\n");
             return sb.ToString();
@@ -57,7 +57,7 @@ namespace HuaweiCloud.SDK.Rds.V3.Model
         public bool Equals(ListDbAgentJobHistorysResponse input)
         {
             if (input == null) return false;
-            if (this.Historys != input.Historys || (this.Historys != null && input.Historys != null && !this.Historys.SequenceEqual(input.Historys))) return false;
+            if (this.Histories != input.Histories || (this.Histories != null && input.Histories != null && !this.Histories.SequenceEqual(input.Histories))) return false;
             if (this.TotalCount != input.TotalCount || (this.TotalCount != null && !this.TotalCount.Equals(input.TotalCount))) return false;
 
             return true;
@@ -71,7 +71,7 @@ namespace HuaweiCloud.SDK.Rds.V3.Model
             unchecked // Overflow is fine, just wrap
             {
                 var hashCode = 41;
-                if (this.Historys != null) hashCode = hashCode * 59 + this.Historys.GetHashCode();
+                if (this.Histories != null) hashCode = hashCode * 59 + this.Histories.GetHashCode();
                 if (this.TotalCount != null) hashCode = hashCode * 59 + this.TotalCount.GetHashCode();
                 return hashCode;
             }