+++ Services/Rds/V3/Model/ValidateInstanceConnectionResponse.cs
@@ -16,7 +16,56 @@ namespace HuaweiCloud.SDK.Rds.V3.Model
     public class ValidateInstanceConnectionResponse : SdkResponse
     {
 
+        /// <summary>
+        /// 响应结果
+        /// </summary>
+        [JsonProperty("resp", NullValueHandling = NullValueHandling.Ignore)]
+        public string Resp { get; set; }
 
 
+
+        /// <summary>
+        /// Get the string
+        /// </summary>
+        public override string ToString()
+        {
+            var sb = new StringBuilder();
+            sb.Append("class ValidateInstanceConnectionResponse {\n");
+            sb.Append("  resp: ").Append(Resp).Append("\n");
+            sb.Append("}\n");
+            return sb.ToString();
+        }
+
+        /// <summary>
+        /// Returns true if objects are equal
+        /// </summary>
+        public override bool Equals(object input)
+        {
+            return this.Equals(input as ValidateInstanceConnectionResponse);
+        }
+
+        /// <summary>
+        /// Returns true if objects are equal
+        /// </summary>
+        public bool Equals(ValidateInstanceConnectionResponse input)
+        {
+            if (input == null) return false;
+            if (this.Resp != input.Resp || (this.Resp != null && !this.Resp.Equals(input.Resp))) return false;
+
+            return true;
+        }
+
+        /// <summary>
+        /// Get hash code
+        /// </summary>
+        public override int GetHashCode()
+        {
+            unchecked // Overflow is fine, just wrap
+            {
+                var hashCode = 41;
+                if (this.Resp != null) hashCode = hashCode * 59 + this.Resp.GetHashCode();
+                return hashCode;
+            }
+        }
     }
 }